requirements.txt
appdirs==1.4.2
Django==1.10.5
django-rest-swagger==0.3.10
djangorestframework==3.5.4
packaging==16.8
path.py==10.1
pkg-resources==0.0.0
pyparsing==2.1.10
PyYAML==3.12
six==1.10.0
tenv==0.1.0
